+++ packages/core/.storybook/src/theme-changer/decorators.jsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { ModalProvider } from '../../../components/Modal';
+import { ModalProvider } from '../../../components/Modal/ModalProvider';
 
 export const withModal = (Story, context) => (
   <ModalProvider>
