--- packages/core/components/Alert/index.ts
@@ -1,3 +1,0 @@
-import { Alert } from './Alert';
-
-export default Alert;
