--- packages/core/components/Avatar/index.ts
@@ -1,3 +1,0 @@
-import { Avatar } from './Avatar';
-
-export default Avatar;
