--- packages/core/components/Checkbox/index.ts
@@ -1,3 +1,0 @@
-import { Checkbox } from './Checkbox';
-
-export default Checkbox;
