--- packages/core/components/Cursor/index.tsx
@@ -1,3 +1,0 @@
-import { Cursor } from './Cursor.css';
-
-export default Cursor;
