--- packages/core/components/Dropdown/index.ts
@@ -1,3 +1,0 @@
-import { Dropdown } from './Dropdown';
-
-export default Dropdown;
