--- packages/core/components/Fieldset/index.ts
@@ -1,3 +1,0 @@
-import { Fieldset } from './Fieldset';
-
-export default Fieldset;
