--- packages/core/components/Frame/index.ts
@@ -1,3 +1,0 @@
-import { Frame } from './Frame';
-
-export default Frame;
