--- packages/core/components/Input/index.ts
@@ -1,3 +1,0 @@
-import { Input } from './Input';
-
-export default Input;
