--- packages/core/components/List/index.ts
@@ -1,3 +1,0 @@
-import { List } from './List';
-
-export default List;
