--- packages/core/components/Modal/index.ts
@@ -1,5 +1,0 @@
-import { Modal } from './Modal';
-import ModalContext from './ModalContext';
-import ModalProvider from './ModalProvider';
-
-export { Modal, ModalContext, ModalProvider };
