--- packages/core/components/RadioButton/index.ts
@@ -1,3 +1,0 @@
-import { RadioButton } from './RadioButton';
-
-export default RadioButton;
