--- packages/core/components/Range/index.ts
@@ -1,3 +1,0 @@
-import { Range } from './Range';
-
-export default Range;
