--- packages/core/components/Tabs/index.ts
@@ -1,4 +1,0 @@
-import { Tab } from './Tab';
-import { Tabs } from './Tabs';
-
-export { Tab, Tabs };
