--- packages/core/components/TaskBar/index.ts
@@ -1,3 +1,0 @@
-import { TaskBar } from './TaskBar';
-
-export default TaskBar;
