--- packages/core/components/TitleBar/index.tsx
@@ -1,3 +1,0 @@
-import { TitleBar } from './TitleBar';
-
-export default TitleBar;
