--- packages/core/components/Tooltip/index.ts
@@ -1,3 +1,0 @@
-import { Tooltip } from './Tooltip';
-
-export default Tooltip;
