--- packages/core/components/Tree/index.ts
@@ -1,3 +1,0 @@
-import { Tree } from './Tree';
-
-export default Tree;
