--- packages/core/components/Video/index.ts
@@ -1,3 +1,0 @@
-import { Video } from './Video';
-
-export default Video;
