+++ packages/core/components/index.ts
@@ -1,25 +1,28 @@
-import { Modal, ModalContext } from './Modal';
-import { Tab, Tabs } from './Tabs';
-import Alert from './Alert';
-import Avatar from './Avatar';
-import Button from './Button';
-import Checkbox from './Checkbox';
-import Cursor from './Cursor';
-import Dropdown from './Dropdown';
-import Fieldset from './Fieldset';
-import Frame from './Frame';
+import { Modal } from './Modal/Modal';
+import { ModalContext } from './Modal/ModalContext';
+import { ModalProvider } from './Modal/ModalProvider';
+import { Tabs } from './Tabs/Tabs';
+import { Tab } from './Tabs/Tab';
+import { Alert } from './Alert/Alert';
+import { Avatar } from './Avatar/Avatar';
+import { Button } from './Button/Button';
+import { Checkbox } from './Checkbox/Checkbox';
+import { Cursor } from './Cursor/Cursor.css';
+import { Dropdown } from './Dropdown/Dropdown';
+import { Fieldset } from './Fieldset/Fieldset';
+import { Frame } from './Frame/Frame';
 import * as GlobalStyle from './GlobalStyle/GlobalStyle.css';
-import Input from './Input';
-import List from './List';
-import ProgressBar from './ProgressBar';
-import RadioButton from './RadioButton';
-import Range from './Range';
-import TaskBar from './TaskBar';
-import TextArea from './TextArea';
-import Tree from './Tree';
-import Tooltip from './Tooltip';
-import TitleBar from './TitleBar';
-import Video from './Video';
+import { Input } from './Input/Input';
+import { List } from './List/List';
+import { ProgressBar } from './ProgressBar/ProgressBar';
+import { RadioButton } from './RadioButton/RadioButton';
+import { Range } from './Range/Range';
+import { TaskBar } from './TaskBar/TaskBar';
+import { TextArea } from './TextArea/TextArea';
+import { Tree } from './Tree/Tree';
+import { Tooltip } from './Tooltip/Tooltip';
+import { TitleBar } from './TitleBar/TitleBar';
+import { Video } from './Video/Video';
 import * as themes from './themes/all';
 
 export {
@@ -36,6 +39,7 @@ export {
   List,
   Modal,
   ModalContext,
+  ModalProvider,
   ProgressBar,
   RadioButton,
   Range,
